Code cleanup (#399)
* rename variables

* rename make target

The name `install-nats-module` is a little misleading, since this target does not work with modules, but with the CR and CRD of the NATS-manager.

* remove unused make targets

A lot of make targets have become obsolete since a release of eventing-manager or nats-manger does no longer generate a module-template. For that reason, all targets that work with modules can be removed.

* remove module specific secitons in the readme

* revert changes for make target name

* make  variables to consts